banana3d-server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
decorator==5.2.1
executing==2.2.1
fal-client==0.7.0
flask==3.1.2
flask-cors==6.0.1
google-auth==2.40.3
google-genai==1.33.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
ipython==9.5.0
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
markupsafe==3.0.2
matplotlib-inline==0.1.7
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
priority==2.0.0
prompt-toolkit==3.0.52
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.1.2
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
werkzeug==3.1.3
wsproto==1.2.0
